backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-socketio==5.13.0
python-multipart==0.0.20
pymongo==4.14.1
python-dotenv==1.1.1
bcrypt==4.3.0
PyJWT==2.8.0
pydantic==2.11.7
httpx==0.28.1
openai==1.106.1
